# src/main/java/com/musala/drone/model/Checks.java
package com.musala.drone.model;

import javax.persistence.*;
import java.time.LocalDateTime;

@Entity
@Table(name = "checks")
public class Checks {

	@Id
	@GeneratedValue(strategy = GenerationType.AUTO)
	@Column(unique = true, nullable = false)
	private Integer id;

	@Column
	private String description;

	@Column
	private LocalDateTime date;

	public Integer getId() {
		return id;
	}

	public void setId(Integer id) {
		this.id = id;
	}

	public String getDescription() {
		return description;
	}

	public void setDescription(String description) {
		this.description = description;
	}

	public LocalDateTime getDate() {
		return date;
	}

	public void setDate(LocalDateTime date) {
		this.date = date;
	}

}